+++ ffi/src/saltyrtc_client_ffi.rs
@@ -13,13 +13,16 @@
 //! re-export all relevant types and generate their own header files.
 
 use std::boxed::Box;
+use std::ffi::CString;
 use std::ptr;
 use std::slice;
 use std::sync::Mutex;
 
-use log::LevelFilter;
+use anyhow::Context;
+use libc::c_char;
+use log::{Level, LevelFilter};
 use log4rs::{Handle as LogHandle, init_config};
-use log4rs::append::console::ConsoleAppender;
+use log4rs::append::{Append, console::ConsoleAppender};
 use log4rs::config::{Appender, Config, Logger, Root};
 use log4rs::encode::pattern::PatternEncoder;
 use saltyrtc_client::crypto::{KeyPair, PrivateKey};
@@ -55,23 +58,88 @@ lazy_static! {
     static ref LOG_HANDLE: Mutex<Option<LogHandle>> = Mutex::new(None);
 }
 
-fn make_log_config(level: LevelFilter) -> Result<Config, String> {
+fn u8_to_levelfilter(level: u8) -> Option<LevelFilter> {
+    Some(match level {
+        LEVEL_TRACE => LevelFilter::Trace,
+        LEVEL_DEBUG => LevelFilter::Debug,
+        LEVEL_INFO => LevelFilter::Info,
+        LEVEL_WARN => LevelFilter::Warn,
+        LEVEL_ERROR => LevelFilter::Error,
+        LEVEL_OFF => LevelFilter::Off,
+        _ => return None
+    })
+}
+
+fn level_to_u8(level: Level) -> u8 {
+    match level {
+        Level::Trace => LEVEL_TRACE,
+        Level::Debug => LEVEL_DEBUG,
+        Level::Info => LEVEL_INFO,
+        Level::Warn => LEVEL_WARN,
+        Level::Error => LEVEL_ERROR,
+    }
+}
+
+pub type LogFunction = unsafe extern "C" fn(level: u8, target: *const c_char, message: *const c_char);
+
+enum LogConfig {
+    Console(LevelFilter),
+    Callback(LogFunction, LevelFilter),
+}
+
+#[derive(Debug)]
+struct CallbackAppender {
+    callback: LogFunction,
+}
+
+impl CallbackAppender {
+    pub fn new(callback: LogFunction) -> Self {
+        Self { callback }
+    }
+}
+
+impl Append for CallbackAppender {
+    fn append(&self, record: &log::Record<'_>) -> anyhow::Result<()> {
+        let target = CString::new(record.target())
+            .context("Could not convert record target to a CString")?;
+        let message = CString::new(record.args().to_string())
+            .context("Could not convert record message to a CString")?;
+        let callback: LogFunction = self.callback;
+        let level = level_to_u8(record.level());
+        unsafe {
+            callback(level, target.as_ptr(), message.as_ptr());
+        }
+        Ok(())
+    }
+
+    fn flush(&self) {}
+}
+
+fn make_log_config(config: LogConfig) -> Result<Config, String> {
     // Log format
     let format = "{d(%Y-%m-%dT%H:%M:%S%.3f)} [{l:<5}] {m} (({f}:{L})){n}";
 
     // Appender
-    let stdout = ConsoleAppender::builder()
-        .encoder(Box::new(PatternEncoder::new(format)))
-        .build();
+    let (appender, level) = match config {
+        LogConfig::Console(level) => {
+            let console = ConsoleAppender::builder()
+                .encoder(Box::new(PatternEncoder::new(format)))
+                .build();
+            (Box::new(console) as Box<dyn Append>, level)
+        }
+        LogConfig::Callback(func, level) => {
+            (Box::new(CallbackAppender::new(func)) as Box<dyn Append>, level)
+        }
+    };
 
     // Create logging config object
     let config_res = Config::builder()
-        .appender(Appender::builder().build("stdout", Box::new(stdout)))
+        .appender(Appender::builder().build("appender", appender))
         .logger(Logger::builder().build("saltyrtc_client", level))
         .logger(Logger::builder().build("saltyrtc_task_relayed_data", level))
         .logger(Logger::builder().build("saltyrtc_task_relayed_data_ffi", level))
         .logger(Logger::builder().build("websocket", level))
-        .build(Root::builder().appender("stdout").build(LevelFilter::Warn));
+        .build(Root::builder().appender("appender").build(LevelFilter::Warn));
 
     config_res.map_err(|e| format!("Could not make log config: {}", e))
 }
@@ -101,21 +169,16 @@ pub extern "C" fn salty_log_init_console(level: u8) -> bool {
     }
 
     // Log level
-    let level_filter = match level {
-        LEVEL_TRACE => LevelFilter::Trace,
-        LEVEL_DEBUG => LevelFilter::Debug,
-        LEVEL_INFO => LevelFilter::Info,
-        LEVEL_WARN => LevelFilter::Warn,
-        LEVEL_ERROR => LevelFilter::Error,
-        LEVEL_OFF => LevelFilter::Off,
-        _ => {
+    let level_filter = match u8_to_levelfilter(level) {
+        Some(lf) => lf,
+        None => {
             eprintln!("salty_log_init_console: Invalid log level: {}", level);
             return false;
         }
     };
 
     // Config
-    let config = match make_log_config(level_filter) {
+    let config = match make_log_config(LogConfig::Console(level_filter)) {
         Ok(config) => config,
         Err(e) => {
             eprintln!("salty_log_init_console: {}", e);
@@ -151,14 +214,9 @@ pub extern "C" fn salty_log_init_console(level: u8) -> bool {
 #[no_mangle]
 pub extern "C" fn salty_log_change_level_console(level: u8) -> bool {
     // Log level
-    let level_filter = match level {
-        LEVEL_TRACE => LevelFilter::Trace,
-        LEVEL_DEBUG => LevelFilter::Debug,
-        LEVEL_INFO => LevelFilter::Info,
-        LEVEL_WARN => LevelFilter::Warn,
-        LEVEL_ERROR => LevelFilter::Error,
-        LEVEL_OFF => LevelFilter::Off,
-        _ => {
+    let level_filter = match u8_to_levelfilter(level) {
+        Some(lf) => lf,
+        None => {
             eprintln!("salty_log_change_level_console: Invalid log level: {}", level);
             return false;
         }
@@ -178,7 +236,7 @@ pub extern "C" fn salty_log_change_level_console(level: u8) -> bool {
     }
 
     // Config
-    let config = match make_log_config(level_filter) {
+    let config = match make_log_config(LogConfig::Console(level_filter)) {
         Ok(config) => config,
         Err(e) => {
             eprintln!("salty_log_change_level_console: {}", e);
@@ -188,6 +246,67 @@ pub extern "C" fn salty_log_change_level_console(level: u8) -> bool {
 
     // Update handle
     handle_opt.as_mut().unwrap().set_config(config);
+
+    // Success!
+    true
+}
+
+/// Initialize logging with a custom callback function that will be called for every log.
+///
+/// Parameters:
+///     callback:
+///         Pointer to a function with the signature
+///         `(uint8_t level, char* target, char* message)`.
+///     level (uint8_t, copied):
+///         The log level, must be in the range 0 (TRACE) to 5 (OFF).
+///         See `LEVEL_*` constants for reference.
+/// Returns:
+///     A boolean indicating whether logging was setup successfully.
+///     If setting up the logger failed, an error message will be written to stdout.
+#[no_mangle]
+pub extern "C" fn salty_log_init_callback(callback: LogFunction, level: u8) -> bool {
+    // Get access to static log handle
+    let mut handle_opt = match LOG_HANDLE.lock() {
+        Ok(handle_opt) => handle_opt,
+        Err(e) => {
+            eprintln!("salty_log_init_callback: Could not get access to static logger mutex: {}", e);
+            return false;
+        }
+    };
+    if handle_opt.is_some() {
+        eprintln!("salty_log_init_callback: A logger is already initialized");
+        return false;
+    }
+
+    // Log level
+    let level_filter = match u8_to_levelfilter(level) {
+        Some(lf) => lf,
+        None => {
+            eprintln!("salty_log_init_callback: Invalid log level: {}", level);
+            return false;
+        }
+    };
+
+    // Config
+    let config = match make_log_config(LogConfig::Callback(callback, level_filter)) {
+        Ok(config) => config,
+        Err(e) => {
+            eprintln!("salty_log_init_callback: {}", e);
+            return false;
+        }
+    };
+
+    // Initialize logger
+    let handle = match init_config(config) {
+        Ok(handle) => handle,
+        Err(e) => {
+            eprintln!("salty_log_init_callback: Could not initialize logger: {}", e);
+            return false;
+        }
+    };
+
+    // Update static logger instance
+    *handle_opt = Some(handle);
 
     // Success!
     true